+++ packages/auth-construct/src/construct.test.ts
@@ -243,7 +243,8 @@ void describe('Auth construct', () => {
   void it('creates email login mechanism with specific settings', () => {
     const app = new App();
     const stack = new Stack(app);
-    const emailBodyFunction = (code: string) => `custom email body ${code}`;
+    const emailBodyFunction = (code: () => string) =>
+      `custom email body ${code()}`;
     const expectedEmailMessage = 'custom email body {####}';
     const customEmailVerificationSubject = 'custom subject';
     new AmplifyAuth(stack, 'test', {
@@ -271,7 +272,8 @@ void describe('Auth construct', () => {
   void it('creates email login mechanism with MFA', () => {
     const app = new App();
     const stack = new Stack(app);
-    const emailBodyFunction = (code: string) => `custom email body ${code}`;
+    const emailBodyFunction = (code: () => string) =>
+      `custom email body ${code()}`;
     const expectedEmailMessage = 'custom email body {####}';
     const customEmailVerificationSubject = 'custom subject';
     const smsVerificationMessageFunction = (code: string) =>
@@ -363,8 +365,8 @@ void describe('Auth construct', () => {
   void it('does not throw if valid email verification message for LINK', () => {
     const app = new App();
     const stack = new Stack(app);
-    const emailBodyFunction = (link: string) =>
-      `valid message ${link} with link`;
+    const emailBodyFunction = (link: (text?: string) => string) =>
+      `valid message ${link()} with link`;
     const customEmailVerificationSubject = 'custom subject';
     assert.doesNotThrow(
       () =>
@@ -378,6 +380,37 @@ void describe('Auth construct', () => {
           },
         })
     );
+    const template = Template.fromStack(stack);
+    template.hasResourceProperties('AWS::Cognito::UserPool', {
+      VerificationMessageTemplate: {
+        DefaultEmailOption: 'CONFIRM_WITH_LINK',
+        EmailMessageByLink: 'valid message {##Verify Email##} with link',
+      },
+    });
+  });
+
+  void it('correctly formats email verification message for LINK with custom link text', () => {
+    const app = new App();
+    const stack = new Stack(app);
+    const emailBodyFunction = (link: (text?: string) => string) =>
+      `valid message ${link('my custom link')} with link`;
+    const customEmailVerificationSubject = 'custom subject';
+    new AmplifyAuth(stack, 'test', {
+      loginWith: {
+        email: {
+          verificationEmailBody: emailBodyFunction,
+          verificationEmailStyle: 'LINK',
+          verificationEmailSubject: customEmailVerificationSubject,
+        },
+      },
+    });
+    const template = Template.fromStack(stack);
+    template.hasResourceProperties('AWS::Cognito::UserPool', {
+      VerificationMessageTemplate: {
+        DefaultEmailOption: 'CONFIRM_WITH_LINK',
+        EmailMessageByLink: 'valid message {##my custom link##} with link',
+      },
+    });
   });
 
   void it('throws error if invalid sms verification message', () => {
@@ -577,9 +610,9 @@ void describe('Auth construct', () => {
               DEFAULTS.PASSWORD_POLICY.minLength.toString(),
             passwordPolicyRequirements:
               defaultPasswordPolicyCharacterRequirements,
-            signupAttributes: '["EMAIL"]',
-            verificationMechanisms: '["EMAIL"]',
-            usernameAttributes: '["EMAIL"]',
+            signupAttributes: '["email"]',
+            verificationMechanisms: '["email"]',
+            usernameAttributes: '["email"]',
             allowUnauthenticatedIdentities: 'true',
           },
         },
@@ -654,9 +687,9 @@ void describe('Auth construct', () => {
                 DEFAULTS.PASSWORD_POLICY.minLength.toString(),
               passwordPolicyRequirements:
                 defaultPasswordPolicyCharacterRequirements,
-              signupAttributes: '["EMAIL"]',
-              verificationMechanisms: '["EMAIL"]',
-              usernameAttributes: '["EMAIL"]',
+              signupAttributes: '["email"]',
+              verificationMechanisms: '["email"]',
+              usernameAttributes: '["email"]',
               googleClientId: 'googleClientId',
               oauthClientId: expectedWebClientId, // same thing
               oauthDomain: `test-prefix.auth.${expectedRegion}.amazoncognito.com`,
@@ -706,7 +739,7 @@ void describe('Auth construct', () => {
 
       assert.equal(
         payload.signupAttributes,
-        '["EMAIL","PHONE_NUMBER","ADDRESS"]'
+        '["email","phone_number","address"]'
       );
     });
 
@@ -720,8 +753,8 @@ void describe('Auth construct', () => {
       });
       const { payload } = storeOutputMock.mock.calls[0].arguments[1];
 
-      assert.equal(payload.usernameAttributes, '["EMAIL","PHONE_NUMBER"]');
-      assert.equal(payload.verificationMechanisms, '["EMAIL","PHONE"]');
+      assert.equal(payload.usernameAttributes, '["email","phone_number"]');
+      assert.equal(payload.verificationMechanisms, '["email","phone_number"]');
     });
 
     void it('stores output when no storage strategy is injected', () => {