+++ packages/backend-storage/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @aws-amplify/backend-storage
 
+## 0.6.0-beta.6
+
+### Minor Changes
+
+- 95a040b: Change storage access path validation from enforcing leading slash to enforcing no leading slash
+
+### Patch Changes
+
+- @aws-amplify/backend-output-storage@0.4.0-beta.4
+
 ## 0.6.0-beta.5
 
 ### Patch Changes