+++ packages/schema-generator/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @aws-amplify/schema-generator
 
+## 0.1.0-beta.4
+
+### Patch Changes
+
+- b14b87e: Include secret name in the generated typescript data schema file
+
+## 0.1.0-beta.3
+
+### Patch Changes
+
+- Updated dependencies [aec89f9]
+- Updated dependencies [2a69684]
+  - @aws-amplify/platform-core@0.5.0-beta.4
+
 ## 0.1.0-beta.2
 
 ### Patch Changes